+++ lib/cuprum/rspec/deferred.rb
@@ -1,0 +1,8 @@
+# frozen_string_literal: true
+
+require 'cuprum/rspec'
+
+module Cuprum::RSpec
+  # Namespace for deferred example groups for validating Cuprum commands.
+  module Deferred; end
+end